common/prud-db/pyproject.toml
[project]
name = "pruddb"
description = "Simple Library used by prud microservices to interact with the central database"
authors = [{ name = "nighmared", email = "joniii@ethz.wtf" }]
version = "1.0.2"
requires-python = ">=3.10"
dependencies = ["argon2-cffi==23.1.*", "sqlmodel", "loguru"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/Nighmared/prud"
